feat(index.html): update content to reflect new versions
The blog link has been removed as the blog was never used and in
place is now the red bubble shop link.

The Explore Beta has been replaced with Explore Features as Tilia
is not in beta so it makes little sense.

The Unity Asset Store blurb and GitHub blurb has been swapped in
position and the Unity Asset Store link is now for the new importer.